test(122488): :white_check_mark: ajusta testes unitários
